# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"John Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/test/countries":{"get":{"tags":["test-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/test/countries/{id}":{"get":{"tags":["test-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/limits":{"get":{"tags":["country-controller"],"operationId":"getAllLimits","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Limit"}}}}}}}},"/api/v1/countries":{"get":{"tags":["country-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}}},"components":{"schemas":{"Continent":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"bonus":{"type":"integer","format":"int32"}}},"Country":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"continent":{"$ref":"#/components/schemas/Continent"},"adjacentCountries":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Limit":{"type":"object","properties":{"idCountryA":{"type":"integer","format":"int64"},"idCountryB":{"type":"integer","format":"int64"}}}}}}